move version back to tscore
