+++ include/iocore/net/Net.h
@@ -40,7 +40,7 @@
  */
 #pragma once
 
-#include "api/Version.h"
+#include "tscore/Version.h"
 #include "iocore/eventsystem/EventSystem.h"
 #include <netinet/in.h>
