+++ src/traffic_logstats/logstats.cc
@@ -25,7 +25,7 @@
 #include "tscore/ink_platform.h"
 #include "tscore/ink_file.h"
 #include "tscore/Layout.h"
-#include "api/Version.h"
+#include "tscore/Version.h"
 #include "tscore/HashFNV.h"
 #include "tscore/ink_args.h"
 #include "tscore/MatcherUtils.h"
