+++ src/traffic_server/Crash.cc
@@ -25,7 +25,7 @@
 #include "swoc/swoc_file.h"
 #include "records/RecCore.h"
 #include "tscore/Layout.h"
-#include "api/Version.h"
+#include "tscore/Version.h"
 #include "tscore/signals.h"
 
 // ucontext.h is deprecated on Darwin, and we really only need it on Linux, so only
