+++ src/traffic_top/traffic_top.cc
@@ -55,7 +55,7 @@
 
 #include "tscore/Layout.h"
 #include "tscore/ink_args.h"
-#include "api/Version.h"
+#include "tscore/Version.h"
 #include "tscore/runroot.h"
 
 using namespace std;
