+++ src/tscore/Version.cc
@@ -22,7 +22,7 @@
  */
 
 #include "tscore/ink_platform.h"
-#include "api/Version.h"
+#include "tscore/Version.h"
 #include "tscore/ink_string.h"
 #include "tscore/ink_file.h"
